+++ src/entities/NonAnimatedRegions.cpp
@@ -179,8 +179,8 @@ void NonAnimatedRegions::update() {
 
   for (const auto& kvp : optimized_tiles_surfaces) {
     const int cell_index = kvp.first;
-    const int row = cell_index / non_animated_tiles.get_num_columns();
-    const int column = cell_index % non_animated_tiles.get_num_columns();
+    const int row = static_cast<int>(cell_index / non_animated_tiles.get_num_columns());
+    const int column = static_cast<int>(cell_index % non_animated_tiles.get_num_columns());
     if (column < column1 || column > column2 || row < row1 || row > row2) {
       indexes_to_clear.push_back(cell_index);
     }
@@ -202,8 +202,8 @@ void NonAnimatedRegions::draw_on_map() {
   }
 
   // Check all grid cells that overlap the camera.
-  const int num_rows = non_animated_tiles.get_num_rows();
-  const int num_columns = non_animated_tiles.get_num_columns();
+  const int num_rows = static_cast<int>(non_animated_tiles.get_num_rows());
+  const int num_columns = static_cast<int>(non_animated_tiles.get_num_columns());
   const Size& cell_size = non_animated_tiles.get_cell_size();
   const Rectangle& camera_position = camera->get_bounding_box();
 
@@ -261,8 +261,8 @@ void NonAnimatedRegions::build_cell(int cell_index) {
       "This cell is already built"
   );
 
-  const int row = cell_index / non_animated_tiles.get_num_columns();
-  const int column = cell_index % non_animated_tiles.get_num_columns();
+  const int row = static_cast<int>(cell_index / non_animated_tiles.get_num_columns());
+  const int column = static_cast<int>(cell_index % non_animated_tiles.get_num_columns());
 
   // Position of this cell on the map.
   const Size cell_size = non_animated_tiles.get_cell_size();
